added standalone server program and library
