+++ openoffload/cpp/framework/include/opof_serverlib.h
@@ -22,9 +22,11 @@
 
 //sessionTable_t * opof_create_server(const char * host, unsigned int port, const char *public_key);
 //
- int opof_add_session_server(sessionRequest_t *parameters, addSessionResponse_t *response);
- int opof_get_session_server(unsigned long sessionId, sessionResponse_t *response);
- int opof_del_session_server(unsigned long sessionId, sessionResponse_t *response);
- sessionResponse_t **opof_get_closed_sessions_server(statisticsRequestArgs_t *request, int *sessionCount);
+//void opof_server(const char *address, unsigned short port, const char* cert, const char* key);
+
+int opof_add_session_server(sessionRequest_t *parameters, addSessionResponse_t *response);
+int opof_get_session_server(unsigned long sessionId, sessionResponse_t *response);
+int opof_del_session_server(unsigned long sessionId, sessionResponse_t *response);
+sessionResponse_t **opof_get_closed_sessions_server(statisticsRequestArgs_t *request, int *sessionCount);
 
 #endif
