+++ openoffload/cpp/framework/include/opof_util.h
@@ -36,4 +36,7 @@ void display_session_response(sessionResponse_t *response);
 void display_session_request(sessionRequest_t *request);
 void convertSessionRequest2cpp(sessionRequest_t *request_c, sessionRequest *request);
 void convertSessionResponse2c(sessionResponse *responsecpp, sessionResponse_t *responsec);
+void convertSessionResponse2cpp(sessionResponse *responsecpp, sessionResponse_t *responsec);
+void convertSessionRequest2c(sessionRequest request, sessionRequest_t *request_c);
+  
 #endif
